mre/requirements.txt
flask==2.2.3
aspose-words==23.2.0